+++ R/exampletestr.R
@@ -1,4 +1,4 @@
-#' @importFrom magrittr '%>%' '%T>%'
+#' @importFrom magrittr '%>%' '%T>%' '%<>%'
 #' @importFrom roxygen2 roclet
 #' @importFrom utils globalVariables
 #' @importFrom devtools use_cran_badge
@@ -6,5 +6,5 @@ NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
 if (getRversion() >= "2.15.1") {
-  globalVariables(c("."))
+  globalVariables(c(".", "file.edit"))
 }